add product add functionality

+++ labb2/Product.java
@@ -1,4 +1,6 @@
 package se.iths.labborationer.labb2;
 
-public record Product(ProductCategory category, String product, int price, int productNumber) {
+import java.math.BigDecimal;
+
+public record Product(ProductCategory category, String product, BigDecimal price, int productNumber) {
 }
